feat: wip working-ish version of the start command
