+++ linkup-cli/src/reset.rs
@@ -1,13 +1,12 @@
 use crate::{
-    background_booting::{BackgroundServices, LocalBackgroundServices},
-    local_config::LocalState,
+    // local_config::LocalState,
     // stop::shutdown,
     CliError,
 };
 
 pub fn reset() -> Result<(), CliError> {
     // Ensure there is some kind of state from before, otherwise reset doesn't make sense
-    let state = LocalState::load()?;
+    // let state = LocalState::load()?;
 
     // shutdown()?;
     // let background_service = LocalBackgroundServices;
